Make compatible with git-pair
Sadly, config.entries uses a regular expression that finds [git-pair] an
acceptable match for pair and authors an acceptable match for
author and thus reads from the wrong list in one's .gitconfig, if one
has a [git-pair] stanza.

A simple and delightfully *✹ fruity *܀ - if perhaps temporary - solution has been found.

Co-authored-by: Daniel Ferguson <daniel.ferguson@futurelearn.com>