Introduce GitConfig::with_config_path add tests
Previously it was unpleasant testing GitConfig's implementation of Store
- it would read & write from the global git config file. By making it
possible to pass in a path to a custom git config file it becomes much
easier. In test code we create a temp file to test against, and
initialize a GitConfig struct with that temp file's path.

The current test suite is extremely minimal and only exercises the happy
path, however it's a start.