Switch git2 from tracking master to release 0.9
Previously we were using the master branch to get access to
GitConfig::remove_multivar - however that method addition has now been
rolled into a release, we can use that instead!

We've shuffled initialisation GitConfig initialisation around a bit in
our test suite - to occur after the git config file is written rather
than before. This is necessary due to a change in the way the libgit2
entries iterator works[1]. Previously it would re-parse the config file
each time it was called, now (since the linked commit) it does *not*
re-parse the config on each call.

[1]: https://github.com/libgit2/libgit2/commit/b78f4ab082030bc02ee2e1a51eca441bfb6f1e8f